% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdd-taxon-info.R
\name{as.list.mdd_taxon_info}
\alias{as.list.mdd_taxon_info}
\title{Convert mdd_taxon_info to list}
\usage{
\method{as.list}{mdd_taxon_info}(x, ...)
}
\arguments{
\item{x}{An object of class \code{mdd_taxon_info}.}

\item{...}{Additional arguments.}
}
\value{
A list representation of the \code{mdd_taxon_info} object.
}
\description{
Convert mdd_taxon_info to list
}
