% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdd-data.R
\name{mdd_cache_dir}
\alias{mdd_cache_dir}
\title{Return the default rmdd cache directory}
\usage{
mdd_cache_dir(path = NULL)
}
\arguments{
\item{path}{Optional path. If \code{NULL}, a user cache directory is constructed.}
}
\value{
A scalar character path.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

This function is stable and its interface is expected to remain compatible.
}
\examples{
mdd_cache_dir(tempdir())
}
