% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdd-data.R
\name{mdd_download}
\alias{mdd_download}
\title{Download a Mammal Diversity Database file}
\usage{
mdd_download(url, dest_dir = mdd_cache_dir(), dest_file = NULL, mode = "wb")
}
\arguments{
\item{url}{URL to the source file.}

\item{dest_dir}{Destination directory for the downloaded file.}

\item{dest_file}{Optional output filename.}

\item{mode}{Transfer mode passed to \code{\link[utils:download.file]{utils::download.file()}}.}
}
\value{
The downloaded file path, invisibly.
}
\description{
Download a Mammal Diversity Database file
}
\examples{
if (interactive()) {
  mdd_download(
    url = "https://www.mammaldiversity.org/robots.txt",
    dest_dir = tempdir()
  )
}
}
