% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdd-data.R
\name{mdd_load}
\alias{mdd_load}
\title{Load an MDD comma-separated export}
\usage{
mdd_load(path)
}
\arguments{
\item{path}{Path to a local CSV export from MDD.}
}
\value{
A tibble.
}
\description{
Load an MDD comma-separated export
}
\examples{
sample_path <- system.file("extdata", "mdd_sample.csv", package = "rmdd")
mdd_load(sample_path)
}
