% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdd-data.R
\name{mdd_reference}
\alias{mdd_reference}
\title{Reference metadata for the Mammal Diversity Database}
\usage{
mdd_reference()
}
\value{
A tibble with package-relevant reference metadata.
}
\description{
Reference metadata for the Mammal Diversity Database
}
