% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdd-records.R
\name{mdd_release_records}
\alias{mdd_release_records}
\title{Return release records used by rmdd}
\usage{
mdd_release_records(
  checklist = NULL,
  synonyms = NULL,
  type_specimen_metadata = NULL
)
}
\arguments{
\item{checklist}{Optional checklist data frame.}

\item{synonyms}{Optional synonym data frame.}

\item{type_specimen_metadata}{Optional type specimen metadata table.}
}
\value{
A named list of tibbles.
}
\description{
Return release records used by rmdd
}
\examples{
x <- mdd_release_records()
names(x)
}
