% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdd-records.R
\name{mdd_synonym_records}
\alias{mdd_synonym_records}
\title{Return synonym records in a normalized tibble}
\usage{
mdd_synonym_records(synonyms = NULL)
}
\arguments{
\item{synonyms}{Optional synonym data frame. Defaults to \code{mdd_synonyms}.}
}
\value{
A tibble.
}
\description{
Return synonym records in a normalized tibble
}
\examples{
mdd_synonym_records()
}
