% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdd-data.R
\name{print.mdd_reference}
\alias{print.mdd_reference}
\title{Print mdd_reference object}
\usage{
\method{print}{mdd_reference}(x, ...)
}
\arguments{
\item{x}{An object of class \code{mdd_reference}.}

\item{...}{Additional arguments.}
}
\value{
The \code{mdd_reference} object, invisibly.
}
\description{
Print mdd_reference object
}
