% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdd-taxon-info.R
\name{print.mdd_taxon_info}
\alias{print.mdd_taxon_info}
\title{Print mdd_taxon_info object}
\usage{
\method{print}{mdd_taxon_info}(x, ...)
}
\arguments{
\item{x}{An object of class \code{mdd_taxon_info}.}

\item{...}{Additional arguments.}
}
\value{
The \code{mdd_taxon_info} object, invisibly.
}
\description{
Print mdd_taxon_info object
}
