% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmdd-package.R
\docType{package}
\name{rmdd-package}
\alias{rmdd}
\alias{rmdd-package}
\title{rmdd: Mammal Diversity Database Tools}
\description{
Tools and packaged data for working with the Mammal Diversity Database (MDD)
in R. The package supports mammal name reconciliation, accepted-name
retrieval, taxon summaries, distribution summaries, and distribution maps
derived from the current bundled MDD release.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/PaulESantos/rmdd}
  \item \url{https://paulesantos.github.io/rmdd/}
  \item \url{https://www.mammaldiversity.org/}
  \item Report bugs at \url{https://github.com/PaulESantos/rmdd/issues}
}

}
\author{
\strong{Maintainer}: Paul E. Santos Andrade \email{paulefrens@gmail.com} [copyright holder]

Authors:
\itemize{
  \item Paul E. Santos Andrade \email{paulefrens@gmail.com} [copyright holder]
}

}
\keyword{internal}
