% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdd-data.R
\name{mdd_cache_dir}
\alias{mdd_cache_dir}
\title{Return the default rmdd cache directory}
\usage{
mdd_cache_dir(path = NULL)
}
\arguments{
\item{path}{Optional path. If \code{NULL}, a user cache directory is constructed.}
}
\value{
A scalar character path.
}
\description{
Return the default rmdd cache directory
}
