% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdd-records.R
\name{mdd_checklist_records}
\alias{mdd_checklist_records}
\title{Return checklist records in a normalized tibble}
\usage{
mdd_checklist_records(checklist = NULL)
}
\arguments{
\item{checklist}{Optional checklist data frame. Defaults to \code{mdd_checklist}.}
}
\value{
A tibble.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

This function is stable and its interface is expected to remain compatible.
}
\examples{
mdd_checklist_records()
}
