% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdd-data.R
\name{mdd_load}
\alias{mdd_load}
\title{Load an MDD comma-separated export}
\usage{
mdd_load(path)
}
\arguments{
\item{path}{Path to a local CSV export from MDD.}
}
\value{
A tibble.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

This function is stable and its interface is expected to remain compatible.
}
\examples{
sample_path <- system.file("extdata", "mdd_sample.csv", package = "rmdd")
mdd_load(sample_path)
}
