% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdd-records.R
\name{mdd_synonym_records}
\alias{mdd_synonym_records}
\title{Return synonym records in a normalized tibble}
\usage{
mdd_synonym_records(synonyms = NULL)
}
\arguments{
\item{synonyms}{Optional synonym data frame. Defaults to \code{mdd_synonyms}.}
}
\value{
A tibble.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

This function is stable and its interface is expected to remain compatible.
}
\examples{
mdd_synonym_records()
}
