% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmdd-package.R
\docType{package}
\name{rmdd-package}
\alias{rmdd}
\alias{rmdd-package}
\title{rmdd: Mammal Diversity Database Tools}
\description{
Development scaffold for a tidyverse-style package that resolves mammal
taxonomic names against the Mammal Diversity Database.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/PaulESantos/rmdd}
  \item \url{https://www.mammaldiversity.org/}
  \item Report bugs at \url{https://github.com/PaulESantos/rmdd/issues}
}

}
\author{
\strong{Maintainer}: Paul E. Santos Andrade \email{paulefrens@gmail.com} [copyright holder]

}
\keyword{internal}
